Crate egglog_core_relations

Structs

AtomId

A component of a query consisting of a function and a list of variables or constants

BaseValueId

an identifier for base value types

BaseValuePrinter

A wrapper used to print a base value.

BaseValues

A registry for base value types and functions on them.

Boxed

A newtype wrapper used to implement the BaseValue trait on types not defined in this crate.

CachedPlan

A cached plan for a given rule.

ColumnId

a particular column in a table

ContainerValueId

an identifier for containers

ContainerValues

CounterId

A counter accessible to actions, useful for generating unique Ids.

Database

A collection of tables and indexes over them.

DisplacedTable

A special table backed by a union-find used to efficiently implement egglog-style canonicaliztion.

DisplacedTableWithProvenance

A variant of DisplacedTable that also stores “provenance” information that can be used to generate proofs of equality.

ExecutionState

A handle on a database that may be in the process of running a rule.

ExternalFunctionId

A user-defined operation that can be invoked from a query

Offset

an opaque offset token – used to encode iterations over a table (within a generation). These always start at 0.

OffsetRange

Pool

A shared pool of objects.

PoolSet

Pooled

An owned value of type T that can be returned to a memory pool when it is no longer used.

ProofStep

QueryBuilder

Builder for the “query” portion of the rule.

Row

A row in a table.

RowId

a numeric offset into a table

RuleBuilder

Builder for the “action” portion of the rule.

RuleId

An identifier for a rule in a rule set

RuleSet

A set of rules to run against a Database.

RuleSetBuilder

Builder for a RuleSet.

SortedWritesTable

TableChange

A summary of the kinds of changes that a table underwent after a merge operation.

TableId

a table in the database

TableSpec

TableVersion

The version of a table.

TaggedRowBuffer

A TaggedRowBuffer wraps a RowBuffer but also keeps track of a source RowId for the row it contains. This makes it useful for materializing the contents of a Subset of a table.

TupleIndex

A mapping from keys to subsets of rows.

Value

A generic identifier representing an egglog value

Variable

a variable in a query

WrappedTable

A WrappedTable takes a Table and extends it with a number of helpful, object-safe methods for accessing a table.

Enums

Constraint

A constraint on the values within a row.

MergeVal

A value that can be written to the database during a merge action.

PlanStrategy

The algorithm used to produce a join plan.

ProofReason

QueryEntry

A representation of a value within a query or rule.

QueryError

Subset

Either or an offset range or a sorted offset vector.

SubsetRef

WriteVal

A value that can be written to a table in an action.

Traits

BaseValue

A simple data type that can be interned in a database.

ContainerValue

A trait implemented by container types.

ExternalFunction

External functions allow external callers to manipulate database state in near-arbitrary ways.

MutationBuffer

A trait specifying a buffer of pending mutations for a Table.

Rebuilder

Custom functions used for tables that encode a bulk value-level rebuild of other tables.

Table

An interface for a table.

Functions

make_external_func

Automatically generate an ExternalFunction implementation from a function.

Type Aliases

MergeFn

The type of closures that are used to merge values in a SortedWritesTable.