egglog_core_relations

Struct DisplacedTable

pub struct DisplacedTable { /* private fields */ }

A special table backed by a union-find used to efficiently implement egglog-style canonicaliztion.

To canonicalize columns, we need to efficiently discover values that have ceased to be canonical. To do that we keep a table of displaced values:

This table has three columns:

(the only key): a value that is no longer canonical in the equivalence relation.

The canonical value of the equivalence class.

The timestamp at which the key stopped being canonical.

We do not store the second value explicitly: instead, we compute it on-the-fly using a union-find data-structure.

This is related to the โLeaderโ encoding in some versions of egglog: Displaced is a version of Leader that only stores ids when they cease to be canonical. Rows are also โautomatically updatedโ with the current leader, rather than requiring the DB to replay history or canonicalize redundant values in the table.

To union new ids l, and r, stage an update Displaced(l, r, ts) where ts is the current timestamp. Note that all tie-breaks and other encoding decisions are made internally, so there may not literally be a row added with this value.

Implementations

impl DisplacedTable

pub fn underlying_uf(&self) -> &UnionFind<Value>