egglog_core_relations

Struct Pooled

pub struct Pooled<T: Clear + InPoolSet<PoolSet>> { /* private fields */ }

An owned value of type T that can be returned to a memory pool when it is no longer used.