egglog_core_relations

Trait BaseValue

pub trait BaseValue:
    Clone
    + Hash
    + Eq
    + Any
    + Debug
    + Send
    + Sync {
    const MAY_UNBOX: bool = false;

    // Provided methods
    fn intern(&self, table: &InternTable<Self, Value>) -> Value { ... }
    fn as_any(&self) -> &dyn Any { ... }
    fn try_box(&self) -> Option<Value> { ... }
    fn try_unbox(_val: Value) -> Option<Self> { ... }
}

A simple data type that can be interned in a database.

Most callers can simply implement this trait on their desired type, with no overrides needed. For types that are particularly small, users can override BaseValue::try_box and BaseValue::try_unbox methods and set BaseValue::MAY_UNBOX to true to allow the Rust value to be stored in-place in a Value.

Regardless, all base value types should be registered in a BaseValues instance using the BaseValues::register_type method before they can be used in the database.

Provided Associated Constants

const MAY_UNBOX: bool = false

Provided Methods

fn intern(&self, table: &InternTable<Self, Value>) -> Value

fn as_any(&self) -> &dyn Any

fn try_box(&self) -> Option<Value>

fn try_unbox(_val: Value) -> Option<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.