egglog_core_relations

Type Alias MergeFn

pub type MergeFn = dyn Fn(&mut ExecutionState<'_>, &[Value], &[Value], &mut Vec<Value>) -> bool + Send + Sync;

The type of closures that are used to merge values in a SortedWritesTable.

The first argument grants access to database using an ExecutionState, the second argument is the current value of the tuple. The third argument is the new, or “incoming” value of the tuple. The fourth argument is a mutable reference to a vector that will be used to store the output of the merge function if it changes the value of the tuple. If it does not, then the merge function should return false.