egglog_core_relations

Struct Boxed

pub struct Boxed<T>(pub T);

A newtype wrapper used to implement the BaseValue trait on types not defined in this crate.

This type is just a helper: users can also implement the BaseValue trait directly on their types if the type is defined in the crate in which the implementation is defined, or if they need custom logic for boxing or unboxing the type.

Tuple Fields
0: T
Implementations

impl<T> Boxed<T>

pub fn new(value: T) -> Self

pub fn into_inner(self) -> T