egglog_core_relations

Struct TaggedRowBuffer

pub struct TaggedRowBuffer { /* private fields */ }

A TaggedRowBuffer wraps a RowBuffer but also keeps track of a source RowId for the row it contains. This makes it useful for materializing the contents of a Subset of a table.

Implementations

impl TaggedRowBuffer

pub fn new(n_columns: usize) -> TaggedRowBuffer

Create a new buffer with the given arity.

pub fn clear(&mut self)

Clear the contents of the buffer.

pub fn is_empty(&self) -> bool

Whether the buffer is empty.

pub fn len(&self) -> usize

The number of rows in the buffer.

pub fn add_row(&mut self, row_id: RowId, row: &[Value]) -> RowId

Add the given row and RowId to the buffer, returning the RowId (in self) for the new row.

pub fn get_row(&self, row: RowId) -> (RowId, &[Value])

Get the row (and the id it was associated with at insertion time) at the offset associated with row.

pub fn get_row_mut(&mut self, row: RowId) -> (RowId, &mut [Value])

pub fn iter(&self) -> impl Iterator<Item = (RowId, &[Value])>

Iterate over the contents of the buffer.

pub fn par_iter(&self) -> impl ParallelIterator<Item = (RowId, &[Value])>

Iterate over the contents of the buffer in parallel.

pub fn non_stale(&self) -> impl Iterator<Item = (RowId, &[Value])>

Iterate over all rows in the buffer, except for the stale ones.

pub fn non_stale_mut(&mut self) -> impl Iterator<Item = (RowId, &mut [Value])>

Iterate over all rows in the buffer, except for the stale ones.

pub fn set_stale(&mut self, row: RowId) -> bool